requirements.txt
spotipy
python-dotenv
psycopg2-binary